Off-by-One fix
This patch adds a test for an extra space at the beginning of lines of code. This specifically addresses https://github.com/styledown/styledown/issues/3, which uses mdextract as a dependency.

In styledown, the syntax requires 3 characters at the beginning of lines (whitespace followed by asterisk followed by whitespace), which throws off the spacing when mdextract finally passes the markdown blocks to Jade.

This patch checks the first line passed to `unpackCode` for a whitespace character in the first position. If it is there, the first character of all lines is sliced off.